get rid of SavedEntityEvent
